Completed TotalWith Tax

+++ src/com/company/TotalWithTax.java
@@ -8,7 +8,7 @@ public class TotalWithTax {
         Scanner scan = new Scanner(System.in);
         System.out.println("Enter the number of price");
         final double TAXRATE=0.0825;
-        double price = scan.nextInt();
+        double price = scan.nextDouble();
         double tax=price*TAXRATE;
         double totalcost=price +tax;
         System.out.println("Price: $"+price);
